functions/process_incoming_image/requirements.txt
google-cloud-firestore==2.3.1
google-cloud-storage==1.42.1
google-cloud-vision==2.4.2
